Moved read/write nb/md utility functions.
